Fix BPTXM bug detecting false re-orgs on instant finality chain (#4768)
It is possible on chains where heads are emitted very quickly that we can be
mid-cycle in the EthConfirmer while a higher head is still in the queue.
This can lead to fetching a receipt with a higher block number than the
currently processing head.

This transaction would be erroneously marked as re-org'd since it isn't
confirmed in the longest seen chain (it is actually confirmed in the future).

The fix is to complete the EthConfirmer receipt checking logic to ignore
transactions confirmed in the future and leave them to be checked on the
next go-around when the next longest chain is popped from the queue, if
necessary.